refactor(tui): split statusview into model, view, and styles files
- Split model.go into three files: model.go (model/update), view.go
  (rendering), styles.go (colors/styles/badge helpers)
- Remove aggregateStatus wrapper, call claude.AggregateStatus directly
- Unify completedDotStyle into dimStyle (identical definitions)
- Remove redundant bg computation in renderSessionBlock (now passed in)
- Convert bg from *lipgloss.Color to lipgloss.Color (never nil)
- Make padLine/emptyLine/renderAccent free functions (no Model receiver)
- Remove unused expandedSessions field and Toggle key handler
- Remove unused cursorStyle
