+++ internal/tui/statusview/model.go
@@ -10,9 +10,6 @@ import (
 	"github.com/charmbracelet/bubbles/help"
 	"github.com/charmbracelet/bubbles/key"
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
-	"github.com/eleonorayaya/utena/internal/claude"
-	"github.com/eleonorayaya/utena/internal/common"
 	"github.com/eleonorayaya/utena/internal/session"
 	"github.com/eleonorayaya/utena/internal/tmux"
 	"github.com/eleonorayaya/utena/internal/tui/provider"
@@ -21,62 +18,11 @@ import (
 
 const collapsedWidth = 14
 
-var (
-	selectedBg = lipgloss.Color("#ffd5d5")
-
-	activeSessionBg   = lipgloss.Color("#f0e8e0")
-	inactiveSessionBg = lipgloss.Color("#e8dfd5")
-
-	accentAttentionColor = lipgloss.Color("#e6537a")
-	accentWorkingColor   = lipgloss.Color("#6a9bc3")
-	accentReviewColor    = lipgloss.Color("#7bb88e")
-	accentIdleColor      = lipgloss.Color("#d8cfc5")
-
-	workspaceStyle = lipgloss.NewStyle().
-			Foreground(lipgloss.Color("#9370b9"))
-
-	cursorStyle = lipgloss.NewStyle().
-			Foreground(lipgloss.Color("#e6537a")).
-			Bold(true)
-
-	sessionNameStyle = lipgloss.NewStyle().
-				Foreground(lipgloss.Color("#4a4a4a"))
-
-	sessionAttachedStyle = lipgloss.NewStyle().
-				Foreground(lipgloss.Color("#2a2a2a")).
-				Bold(true)
-
-	windowActiveStyle = lipgloss.NewStyle().
-				Foreground(lipgloss.Color("#6a9bc3"))
-
-	windowDimStyle = lipgloss.NewStyle().
-			Foreground(lipgloss.Color("#8a7873"))
-
-	dimStyle = lipgloss.NewStyle().
-			Foreground(lipgloss.Color("#8a7873"))
-
-	completedDotStyle = lipgloss.NewStyle().
-				Foreground(lipgloss.Color("#8a7873"))
-
-	attentionBadgeStyle = lipgloss.NewStyle().
-				Bold(true).
-				Foreground(lipgloss.Color("#fef6f0")).
-				Background(lipgloss.Color("#d16577"))
-
-	workingBadgeStyle = lipgloss.NewStyle().
-				Foreground(lipgloss.Color("#d97c6e"))
-
-	reviewBadgeStyle = lipgloss.NewStyle().
-				Foreground(lipgloss.Color("#5fafa5"))
-
-)
-
 type tickMsg time.Time
 
 type Model struct {
 	sessions           []session.Session
 	windowsBySession   map[string][]tmux.Window
-	expandedSessions   map[string]bool
 	currentTmuxSession string
 	paneID             string
 	focused            bool
@@ -101,7 +47,6 @@ func New() Model {
 		paneID:             paneID,
 		currentTmuxSession: currentSession,
 		windowsBySession:   make(map[string][]tmux.Window),
-		expandedSessions:   map[string]bool{currentSession: true},
 	}
 }
 
@@ -218,21 +163,6 @@ func (m Model) onKeyMsg(msg tea.KeyMsg) (Model, tea.Cmd) {
 			return m, provider.ActivateSession(s.ID)
 		}
 		return m, nil
-	case key.Matches(msg, keys.Toggle):
-		if m.cursor < len(ordered) {
-			s := ordered[m.cursor]
-			if s.IsAttached {
-				return m, nil
-			}
-			tmuxName := s.TmuxSessionName
-			if m.expandedSessions[tmuxName] {
-				delete(m.expandedSessions, tmuxName)
-			} else {
-				m.expandedSessions[tmuxName] = true
-				return m, provider.FetchWindows(tmuxName)
-			}
-		}
-		return m, nil
 	case key.Matches(msg, keys.Collapse):
 		return m, nil
 	}
@@ -255,253 +185,4 @@ func (m Model) orderedActiveSessions() []session.Session {
 		return sessionDisplayName(result[i]) < sessionDisplayName(result[j])
 	})
 	return result
-}
-
-func (m Model) View() string {
-	if m.isExpanded() {
-		return m.expandedView()
-	}
-	return m.collapsedView()
-}
-
-func (m Model) collapsedView() string {
-	maxNameLen := m.width - 2
-	if maxNameLen < 1 {
-		maxNameLen = 1
-	}
-
-	var lines []string
-	for _, sess := range m.orderedActiveSessions() {
-		name := sessionDisplayName(sess)
-		name = truncate(name, maxNameLen)
-		lines = append(lines, "  "+name)
-	}
-
-	return strings.Join(lines, "\n")
-}
-
-func (m Model) expandedView() string {
-	innerWidth := m.width
-	if innerWidth < 1 {
-		innerWidth = 1
-	}
-
-	ordered := m.orderedActiveSessions()
-
-	var lines []string
-	for i, s := range ordered {
-		selected := i == m.cursor && m.focused
-		bg := &inactiveSessionBg
-		if selected {
-			bg = &selectedBg
-		} else if s.IsAttached {
-			bg = &activeSessionBg
-		}
-		lines = append(lines, m.emptyLine(innerWidth, bg))
-		lines = append(lines, m.renderSessionBlock(s, i, innerWidth)...)
-		lines = append(lines, m.renderWindows(s, innerWidth, bg)...)
-		lines = append(lines, m.emptyLine(innerWidth, bg))
-	}
-
-	if len(ordered) == 0 {
-		lines = append(lines, dimStyle.Render("  no active sessions"))
-	}
-
-	return strings.Join(lines, "\n")
-}
-
-func (m Model) renderSessionBlock(s session.Session, idx, innerWidth int) []string {
-	selected := idx == m.cursor && m.focused
-	active := s.IsAttached
-
-	bg := &inactiveSessionBg
-	if selected {
-		bg = &selectedBg
-	} else if active {
-		bg = &activeSessionBg
-	}
-
-	name := sessionDisplayName(s)
-	wsName := ""
-	if s.Workspace != nil {
-		wsName = s.Workspace.Name
-	}
-
-	nStyle := sessionNameStyle
-	if active {
-		nStyle = sessionAttachedStyle
-	}
-	wsStyle := workspaceStyle
-	bStyle, badgeText := claudeBadgeParts(s.ClaudeSessions)
-
-	if bg != nil {
-		nStyle = nStyle.Background(*bg)
-		wsStyle = wsStyle.Background(*bg)
-		if bStyle != nil {
-			s := (*bStyle).Background(*bg)
-			bStyle = &s
-		}
-	}
-
-	badgeStr := ""
-	if bStyle != nil {
-		badgeStr = (*bStyle).Render(badgeText)
-	}
-
-	barColor := accentBarColor(s.ClaudeSessions)
-	barBase := lipgloss.NewStyle().Foreground(barColor)
-	accentWidth := 2
-
-	maxName := innerWidth - accentWidth - lipgloss.Width(badgeStr)
-	if badgeStr != "" {
-		maxName--
-	}
-	if maxName < 4 {
-		maxName = 4
-	}
-	truncName := truncate(name, maxName)
-	nameStr := nStyle.Render(truncName)
-
-	accent := renderAccent(barBase, bg)
-	nameLine := accent + nameStr
-	if badgeStr != "" {
-		usedWidth := accentWidth + lipgloss.Width(truncName) + lipgloss.Width(badgeStr)
-		pad := innerWidth - usedWidth
-		if pad < 1 {
-			pad = 1
-		}
-		padStr := strings.Repeat(" ", pad)
-		if bg != nil {
-			padStr = lipgloss.NewStyle().Background(*bg).Render(padStr)
-		}
-		nameLine = accent + nameStr + padStr + badgeStr
-	}
-	nameLine = m.padLine(nameLine, innerWidth, bg)
-
-	var result []string
-	result = append(result, nameLine)
-
-	if wsName != "" {
-		timeStr := ""
-		if !s.LastUsedAt.IsZero() {
-			tStyle := dimStyle
-			if bg != nil {
-				tStyle = tStyle.Background(*bg)
-			}
-			timeStr = tStyle.Render(" · " + common.TimeAgo(s.LastUsedAt))
-		}
-		wsLine := accent + wsStyle.Render(wsName) + timeStr
-		wsLine = m.padLine(wsLine, innerWidth, bg)
-		result = append(result, wsLine)
-	}
-
-	return result
-}
-
-func (m Model) padLine(line string, width int, bg *lipgloss.Color) string {
-	if bg == nil {
-		return line
-	}
-	lineWidth := lipgloss.Width(line)
-	if lineWidth < width {
-		line += lipgloss.NewStyle().Background(*bg).Render(strings.Repeat(" ", width-lineWidth))
-	}
-	return line
-}
-
-func (m Model) emptyLine(width int, bg *lipgloss.Color) string {
-	if bg != nil {
-		return lipgloss.NewStyle().Background(*bg).Render(strings.Repeat(" ", width))
-	}
-	return ""
-}
-
-func renderAccent(barBase lipgloss.Style, bg *lipgloss.Color) string {
-	if bg != nil {
-		return barBase.Background(*bg).Render("▐") + lipgloss.NewStyle().Background(*bg).Render(" ")
-	}
-	return barBase.Render("▐") + " "
-}
-
-func claudeBadgeParts(sessions []claude.ClaudeSession) (*lipgloss.Style, string) {
-	status := aggregateStatus(sessions)
-	switch status {
-	case claude.StatusNeedsAttention:
-		s := attentionBadgeStyle
-		return &s, " ! "
-	case claude.StatusWorking:
-		s := workingBadgeStyle
-		return &s, " ~ "
-	case claude.StatusReadyForReview:
-		s := reviewBadgeStyle
-		return &s, " ✓ "
-	case claude.StatusDone:
-		s := completedDotStyle
-		return &s, " ✓ "
-	default:
-		return nil, ""
-	}
-}
-
-func (m Model) renderWindows(s session.Session, innerWidth int, bg *lipgloss.Color) []string {
-	windows := m.windowsBySession[s.TmuxSessionName]
-	if len(windows) == 0 {
-		return nil
-	}
-
-	barColor := accentBarColor(s.ClaudeSessions)
-	barBase := lipgloss.NewStyle().Foreground(barColor)
-	accent := renderAccent(barBase, bg)
-
-	var lines []string
-	for _, w := range windows {
-		marker := "  "
-		wStyle := windowDimStyle
-		if w.Active {
-			marker = "› "
-			wStyle = windowActiveStyle
-		}
-		if bg != nil {
-			wStyle = wStyle.Background(*bg)
-		}
-		line := accent + wStyle.Render(marker+w.Name)
-		line = m.padLine(line, innerWidth, bg)
-		lines = append(lines, line)
-	}
-	return lines
-}
-
-func accentBarColor(claudeSessions []claude.ClaudeSession) lipgloss.Color {
-	status := aggregateStatus(claudeSessions)
-	switch status {
-	case claude.StatusNeedsAttention:
-		return accentAttentionColor
-	case claude.StatusWorking:
-		return accentWorkingColor
-	case claude.StatusReadyForReview:
-		return accentReviewColor
-	default:
-		return accentIdleColor
-	}
-}
-
-func sessionDisplayName(s session.Session) string {
-	if s.Name != "" {
-		return s.Name
-	}
-	return s.TmuxSessionName
-}
-
-func truncate(s string, maxLen int) string {
-	if len(s) <= maxLen {
-		return s
-	}
-	if maxLen <= 1 {
-		return s[:maxLen]
-	}
-	return s[:maxLen-1] + "…"
-}
-
-func aggregateStatus(sessions []claude.ClaudeSession) claude.ClaudeSessionStatus {
-	return claude.AggregateStatus(sessions)
 }
